+++ services/events/notifications.js
@@ -4,15 +4,7 @@ const mailService = require('../mail/mail-service')
 const firebaseService = require('../firebase/firebase-service')
 const moment = require('moment-timezone')
 const { getTzByLatLng } = require('../../utils/misc/timezone')
-const rangerAPIEnabled = `${process.env.RANGER_API_ENABLED}` === 'true'
-const rangerAPITopic = process.env.RANGER_API_SQS_QUEUE
-let rangerAPIMessageQueue
 let emailTemplate
-
-if (rangerAPIEnabled) {
-  const { MessageQueue } = require('@rfcx/message-queue')
-  rangerAPIMessageQueue = new MessageQueue('sqs', { endpoint: '' })
-}
 
 /**
  * Define html template for email notifications on app initialization
@@ -75,34 +67,10 @@ async function notify (event) {
       latitude: `${event.stream.latitude}`,
       longitude: `${event.stream.longitude}`
     }
-    const promises = [
+    await Promise.all([
       sendEmails(splittedSubs.Email, data),
       sendPushNotifications(splittedSubs['Push Notification'], pnData)
-    ]
-    if (rangerAPIEnabled) {
-      const rangerMessageData = {
-        id: event.id,
-        start: event.start,
-        end: event.end,
-        stream: {
-          id: event.stream.id,
-          name: event.stream.name,
-          latitude: event.stream.latitude,
-          longitude: event.stream.longitude
-        },
-        project: {
-          id: event.stream.project.id,
-          name: event.stream.project.name
-        },
-        classification: {
-          value: event.classification.value,
-          title: event.classification.title
-        },
-        createdAt: event.createdAt
-      }
-      promises.push(sendRangerSQS(rangerMessageData))
-    }
-    await Promise.all(promises)
+    ])
   }
 }
 
@@ -152,14 +120,6 @@ function sendPushNotifications (users, data) {
     body
   }
   return firebaseService.sendToTopic(opts)
-}
-
-/**
- * Sends SQS message with event data to Ranger API
- * @param {*} data
- */
-async function sendRangerSQS (data) {
-  return rangerAPIMessageQueue.publish(rangerAPITopic, data)
 }
 
 module.exports = {