+++ tasks/segment-created/classifier-message-queue/default.js
@@ -5,10 +5,6 @@
  */
 
 const ClassifierMessageQueue = require('.')
-const SQSClient = require('../../../utils/message-queue/sqs-client')
-
-const sqs = new SQSClient({ endpoint: process.env.MESSAGE_QUEUE_ENDPOINT })
-const options = { queuePrefix: 'classifier' }
-const messageQueue = new ClassifierMessageQueue(sqs, options)
+const messageQueue = new ClassifierMessageQueue('sqs', { queuePrefix: 'classifier' })
 
 module.exports = messageQueue