+++ utils/message-queue/sns.js
@@ -1,17 +1,17 @@
 const isEnabled = process.env.MESSAGE_QUEUE_ENABLED === 'true'
 
 /**
- * Get the default MessageQueue from the environment configuration
- * @module defaultMessageQueue
+ * Get the SNS MessageQueue from the environment configuration
+ * @module snsMessageQueue
  * @type {MessageQueue}
  */
 let messageQueue
 
 if (isEnabled) {
   const { MessageQueue } = require('@rfcx/message-queue')
-  const messageQueue = new MessageQueue('sqs', {
+  const messageQueue = new MessageQueue('sns', {
     endpoint: process.env.MESSAGE_QUEUE_ENDPOINT,
-    topicPrefix: process.env.MESSAGE_QUEUE_PREFIX
+    topicPostfix: process.env.NODE_ENV
   })
   messageQueue.isEnabled = () => true
 } else {